examples/pdf_embedding/pyproject.toml
[project]
name = "pdf-embedding"
version = "0.1.0"
description = "Simple example for cocoindex: build embedding index based on local PDF files."
requires-python = ">=3.11"
dependencies = [
    "cocoindex>=0.1.52",
    "python-dotenv>=1.0.1",
    "marker-pdf>=1.5.2",
]

[tool.setuptools]
packages = []
